+++ package.json
@@ -18,8 +18,8 @@
         "sqlite3": "^5.0.0"
     },
     "devDependencies": {
-        "cross-env": "^7.0.2",
-        "eslint": "^7.13.0",
+        "cross-env": "^7.0.3",
+        "eslint": "^7.14.0",
         "jest": "^26.6.3",
         "knex-cleaner": "^1.3.1",
         "supertest": "^6.0.1"
